pyproject.toml
[tool.poetry]
name = "dbmp"
version = "0.1.0"
description = ""
authors = ["iviv62 <ivcho2213@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.3"
gunicorn = "^21.2.0"
psycopg2 = "^2.9.6"
sentence-transformers = "^2.2.2"
farm-haystack = "^1.20.0"
djangorestframework = "^3.14.0"
pypdf2 = "^3.0.1"
python-docx = "^0.8.11"
pinecone-client = "^2.2.2"
django-cors-headers = "^4.2.0"
logger = "^1.4"
poetry = "^1.5.1"
whitenoise ="^6.5.0"
boto3 = "^1.28.36"
django-storages = "^1.13.2"
xhtml2pdf = "^0.2.11"
stubs = "^1.0.0"
fuzzywuzzy = "^0.18.0"
retry = "^0.9.2"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[tool.black]
line-length=79
target-version=["py311"]

[tool.mypy]
exclude = ['venv', '.venv', 'migrations']

[tool.isort]
py_version=311
profile="black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]
production = ["Django", "gunicorn", "psycopg2"]